Increase initial DB fetch for LB overall
LB was fetching the top 100 total XP users, then sorting those by total level. This meant that a user with 200m XP in a few skills would be on the LB over a user with 14m in many skills.

This makes it fetch the top 2000, sort by total level, then return the top 100 of the sorted list.

Effect: Users with lower total XP but higher total level will no longer be incorrectly excluded.